AIN1 and AIN2 renamed to pwm_pin and dir_pin. Part 4 almost done, needs to plot, etc. still!

+++ Project_2/include/timer_usec.h
@@ -4,6 +4,6 @@ class Timer_usec
 {
 public:
     Timer_usec();
-    void init(uint16_t period_us, int duty_cycle);
-    void set_duty_cycle(int duty_cycle);
+    void init(uint16_t period_us, double duty_cycle);
+    void set_duty_cycle(double duty_cycle);
 };
