+++ Project_2/src/timer_usec.cpp
@@ -4,14 +4,15 @@
     
 Timer_usec::Timer_usec(){}
 
-void Timer_usec:: init(uint16_t period_us, int duty_cycle){
+void Timer_usec:: init(uint16_t period_us, double duty_cycle){
     uint32_t ticks = ((uint32_t)period_us * 2) -1;
 
     TCCR0A = 0; // Setting timer1 to normal operation
     TCCR0B = 0;
     TCNT0 = 0; // Initialize counter value
     OCR0A = (uint8_t) ticks;
-    OCR0B = ((uint32_t)OCR0A * duty_cycle) / 100.0;
+    OCR0B = (uint8_t)((double)OCR0A * duty_cycle / 100.0);
+
 
     TCCR0B |= (1<<WGM01);
     TIMSK0 |= (1<<OCIE0A);
@@ -21,6 +22,6 @@ void Timer_usec:: init(uint16_t period_us, int duty_cycle){
 
 }
 
-void Timer_usec :: set_duty_cycle(int duty_cycle) {
-    OCR0B = ((uint32_t)OCR0A * duty_cycle) / 100.0;
+void Timer_usec :: set_duty_cycle(double duty_cycle) {
+    OCR0B = (uint8_t)((double)OCR0A * duty_cycle / 100.0);
 }
